Implement performance bisection task+evaluators
This change implements the top-down logic of the performance bisection
using the tasks+evaluators model used by the execution engine. Here we
implement the initial task graph seeding which follows the following
pattern:

           find_culprit
          /            \
         /              \
        v                v
  SubGraph(CL0)  SubGraph(CLM)

The SubGraph(...) instances above expand to the `read_value` subgraphs
that the singular `find_culprit` node will depend on. This structure
allows us to have the find_culprit node depend on the individual
`read_value` attempts so that we can track each attempt's progress
individually, which will eventually depend on two builds.

In the evaluator, we make the determination of whether we should add
more subgraphs or expand existing subgraphs based on the data we get
from the `read_value` tasks. This allows us to consolidate the
evaluation process into a single module which we can then test by
simulating different `read_value` outcomes we might encounter in the
course of bisecting a performance regression.

We are re-using a lot of the implementation in the job module, which
does this evaluation process in the `Explore()` method in the JobState.

We can follow the same model for implementing A/B testing, functional
bisection, and other potential workflows re-using the `find_culprit`
task graph and applying different evaluators.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:988912
Change-Id: I56a2aaffad0be8e6baf24c682f72108fa6373b4e