[Telemetry] Confirm startup tracing on establishing DevTools conection
When establishing a DevTools connection, the TracingBackend is also
created.

Therefore, if Chrome startup tracing has already be enabled, we can
take this chance to send a new Tracing.start command to confirm
and update settings on the tracing session; rather than waiting to
the end close to the StopTracing command.

Gladly this also makes the code a bit more readable and easy to
understand.

Bug: chromium:914092
Change-Id: I59072e437c989375518d94de0e1a5f92f450480c