[Telemetry] Move --verbose option to run_tests
The --verbose option is now handled by Results Processor, so to avoid
duplication it's moved from browser_options to run_tests. Tests run from
Catapult CQ now pass -v option in the same way as telemetry_unittests
in Chromium do.

Also the Popen hook that was introduced for debugging one specific bug
is removed because it's not needed anymore.

Bug: chromium:985712
Change-Id: Ib335ced67ae4f586a8bd93525cfe72512e886180