Prepare to Avoid Means in Change Detection
This change introduces an alternative calculation when doing change
point detection by avoiding computing the means of the values from the
individual attempts. The alternative implementation is there so that we
can first count the number of times we encounter cases where the
alternative computation has a different finding from the current
approach which uses the mean of the results of an attempt.

When this change is in production we can set up a temporary log metric
to catch the number of times the log message appears, and make a
determination on how often the approaches differ -- and whether it's
qualitatively better to use the alternative approach.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:1011273
Change-Id: I0a8b8f485daa863303809ef243f90d8491efd90e