[typ] Create FileManager class which encases logic for handling files
This CL will create an interface for which the Artifacts class can
interact with the files on disc. The API of this file manager will
follow closely the API of blinkpy's file manager. This CL also adds
the optional force_overwrite parameter to CreateArtifact, so that it
can force the function to skip checking if the file exists.

Bug: chromium:982208
Change-Id: I3ed666bcb7edb029d8637e042c9fa2e286766160
TBR=dpranke@chromium.org
R=isamsonov@google.com