Decrease Job Frozen Threshold
Make the frozen job threshold more aggressive. We used to wait 6 hours
of no activity before considering a job is frozen, so now we're reducing
that to 1h for the following reasons:

- Given that we have a scheduler that's able to enforce exclusivity of
the pool now, this means we are potentially making pools idle for 6
hours before we consider any action.

- Retrying jobs are a good idea anyway, if we fail to find/make progress
for a number of times. Calling a job "failed" sooner is better than
potentially waiting too long before letting other jobs run sooner.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:1011728
Change-Id: I3d8b11b9c66a2ff941635dc5e38b6724c91d516c