Do not use Datastore as a cache
This change makes the commit cache only use memcache and the in-memory
cache instead of writing through to Datastore. For the purpose of a
cache, memcache and the in-memory cache should fine without needing to
ensure transactional semantics for the operations on the commit cache.

This change also allows us to reduce the amounts of calls to the gitiles
service when resolving commits by checking the cache first if we've seen
these changes recently.

We can apply this transformation for the other cache-only objects we
have in Datastore.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:1006888
Change-Id: I92864263a6ed2b2a14107f410ed0d640444c030a