[typ] Add dirname method to artifacts FileManager
This method adds the dirname method to the FileManager class.
It also updates the CreateArtifacts method to use this new method
instead of the os.path one. We should make sure that the Artifacts
class is only using methods from the FileManager in order to interact
with files. This CL also adds a verification for duplicate artifacts
added through the AddArtifact method.

r=isamsonov@google.com

Bug: chromium:982208
Change-Id: I19a28d5ec98c9711a74e21d5944ca417f7517403
TBR=dpranke@chromium.org