[Telemetry] Move expectations story skipping into story_filter.py.
Before this change, expectations code was littered around in several
places including benchmark.py, story_runner.py, commands.py, etc.
Each of those modules needed their own unit tests for
expectations-related behavior as well. This change factors out all the
expectations-related code into a single ShouldSkip function in
story_filter.py. story_filter.py comes to have two related
functionalities:
1. Filtering stories that should not be run or marked skipped.
2. Skipping stories that should be skipped.

In the future, we want to support a flag such as
--story=<exact story name> (or maybe --stories=<list of
exact story names>). For that feature, we will need to
augment both 1 and 2 because:
1. We should filter out any story that does not match
that flag.
2. We should probably ignore expectations if someone passes
an exact story name. (Or maybe make this configurable?)

The point of the above is that this story filtering and story skipping
code is related and could benefit from being part of a single module.

Note that this also allows improvements to the unit test
code. In a few places, we can now mock out story_filter instead of
having to build fake expectations. Making changes to story filtering
or to expectations will not require editing any story_runner_unittests.

Bug: chromium:985103
Change-Id: I966e9541b2df1e17f561077d119eea1c179e35d2