[Telemetry] Save measurements as dict
This CL disallows adding multiple measurements with the same name, and
changes the serialization format for measurement to dict instead of list.
This is to make ad-hoc measurements API more strict and unambiguous.

Note that landing this is blocked on resolving the issue with oilpan_gc_times
bechmarks (see crbug.com/1011787) which currently do add several measurements
with the same name.

Bug: chromium:999484
Change-Id: Ib75b58fc0f244e307f5897800264e6eb76b37ba9