PinpointScheduler: Wire the Execution Engine
This is an initial integration attempt to get the execution engine
implementation testable through an opt-in mechanism in the UI. In this
change we do the following:

- Wire up the task graph creation from when we're creating a job through
the UI, to allow us to switch execution towards using the execution
engine.

- Create a single evaluator which wires up all the evaluators required
to evaluate the performance bisection task graph.

- Integrate the task evaluation workflow in the Pub/Sub handler, to
allow us to make progress on a job through the execution engine.

- Implement a set of simple end-to-end tests to cover the new code.

- Implement an evaluator which transforms the task graph information
into something that the current UI can render.

The intent is to incrementally migrate all the workflows to move from
using the quests+execution framework towards the execution engine, to
allow us to make progress in smaller, more predictable chunks of
functionality. It also allows us to more effectively refactor and test
in the wild the use-cases we need to be supporting in Pinpoint.

When all the use-cases are done, we can then remove all traces of the
job state pickling implementation, and build more on top of the
execution engine approach.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:988912
Change-Id: Ib73c79dd1fc3e15f6219a0c1f067a2b565790fd9