[Telemetry] New intermediate file format
This CL implements a new intermediate file format and writes it to an
_intermediate_results.jsonl file. The main difference from the
old format is that testResult messages are now self-sufficient, i.e.
they contain all the information necessary for processing results.
The metadata is now stored as a special artifact of every story run.
The benchmarkRun messages are eliminated.

The old format is still there in the _telemetry_results.jsonl file for
the transitional period.

Bug: chromium:981349
Change-Id: Id86e29b1f3a7f2f57b220c0937db73cdaa09d9a6