Force IPv4 for Telemetry proxy server
Forces Chrome to use IPv4 when connecting to the proxy server set up
by Telemetry. This is done by specifying "127.0.0.1" instead of
"localhost", thus avoiding the automatic expansion into both "::1" and
"127.0.0.1". This avoids issues with the automatically chosen device
port being free for IPv4, but in use by some other process for IPv6,
which ultimately causes Chrome to connect to the other process and
cause tests to fail.

Bug: chromium:1005971
Change-Id: I3fc19253e07d461aa20ca5a8f51a96097dba9b8c