trace-viewer: Make some style changes for the toolbar.
If the trace-file has a long path (which is often the case for traces
generated on the bots from pinpoint/perf bots), the title in the toolbar
gets too big, and pushes the search textfield/buttons on the right
off-screen. Fix this by turning off shrinking for left/right controls,
and overflow-hide the title.

BUG=chromium:13134234234

Change-Id: I7907dfffbc826574cf9a239efe997534faa6abab