[Telemetry] Validate response of Tracing.start request
In startup tracing mode, the response from Tracing.start request was
ignored. In case of failure to start tracing Telemetry would not
know about this until much later after the story run had completed.

In this CL we make it so the response is always validated in all
cases, and add some additional debug information to help diagnose
failures.

Bug: chromium:1011232
Change-Id: I4d7656ce53c8151a8142581994f8ff7561718b5b