[pinpoint] Replace "tir_label" with "grouping_label"
The concept of "tir_label" (a timeline interaction record label) no
longer exists in Telemetry. It was replaced with a simpler
"grouping_label" which just allows grouping stories into meaningful
groups.

This CL updates pinpoint code to use the new name, and be ready for
clients that start using the new name in requests.

Bug: chromium:974237
Change-Id: Ic741e0e2a982ea20aa3e6b11657dcb6f06970526