Update patch set 9
Patch Set 9:

Really cool! Can you please file a crbug and link to that from the CL description? Also, if possible, please upload a trace-file with this change and link to that from the CL description and/or the crbug too.

Patch-set: 9
Reviewer: Gerrit User 1000111 <1000111@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>