Switch back test runner to chromium
Chromium binaries are causing dashboard tests to time out. After
debugging with xvfb turned off, it appears that the entire browser
process freezes for unclear reasons. The logs always mention getting a
proxy config as the last message. This problem does not show up on
chrome builds, so switching it back. Adding the
--enable-features=ForceWebRequestProxyForTest flag as suggested by the
network team to get around the earlier proxy issues we were seeing.

Bug: chromium:1015942
Change-Id: I402fe99e1a80a039186dc21c658a8be9e4485349