Introduces timeout for dashboard dev server tests
Dashboard dev server tests has been flakily timing out recently because the dev
server process hangs. It is unclear why that's happening, but in the meantime,
introduce a 2 minute timeout with one retry. We hope the one retry will increase
the success rate, and if not, then at least test will fail fast as opposed to
waiting for two hours. The root cause of the server hang is being investigated
separately.

Bug: chromium:1015942
Change-Id: Ie5aecae2f94f3f9c0d23b6a499fa3cd49a64df05