Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Is there context for this? Did you hit this on a bot, or running something (tests, I presume?) locally, or...?

Locally. I would imagine bots hit this too, but it likely doesn't matter as I'm sure we blow away the device on every run.
When running the first time through it'll copy to one directory, and the second time through it'll copy to the nested directory. This means that on subsequent runs, if you modify a file it's only updated in the nested directory. If the code is configured to read from the non-nested directory you get stale data.

For example: src=test/data, target=test/data file of interest is 1.html

1. On first run source is copied to test/data/1.html
2. Run again, as target directory exists, end up with test/data/data.
3. Modify 1.html locally.
4. Run again, only test/data/data/1.html is modified on device.

Patch-set: 2