Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > Is there context for this? Did you hit this on a bot, or running something (tests, I presume?) locally, or...?
> > 
> > Locally. I would imagine bots hit this too, but it likely doesn't matter as I'm sure we blow away the device on every run.

we sure do

> > When running the first time through it'll copy to one directory, and the second time through it'll copy to the nested directory. This means that on subsequent runs, if you modify a file it's only updated in the nested directory. If the code is configured to read from the non-nested directory you get stale data.
> > 
> > For example: src=test/data, target=test/data file of interest is 1.html
> > 
> > 1. On first run source is copied to test/data/1.html
> > 2. Run again, as target directory exists, end up with test/data/data.
> > 3. Modify 1.html locally.
> > 4. Run again, only test/data/data/1.html is modified on device.
> 
> And to be clear, by 'it' I mean java instrumentation tests, which use this code.

and this is (i assume) w/ weblayer instrumentation tests? as-is in chromium/src, or w/ a local patch?

While I do think this is probably worth fixing in one way or another, I don't think this is the right layer for the fix. adb_wrapper tries to be a pretty direct interface to adb; higher layers (e.g. device_utils and the android test runner) would be responsible for adapting like this. If adb itself changed behavior (as happened w/ the pre-1.0.36 code), then we'd want it in this layer, but it's not clear that that's the case here. Either way, I'd like to repro.

Patch-set: 2
Reviewer: Gerrit User 1181690 <1181690@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>