Update patch set 2
Patch Set 2:

snip

> Happy to fix this higher in the stack. I suspect adb hasn't really changed here, but I did not attempt to verify that. I believe that would mean code higher in the stack needs to recursively compare all files, and have a special case for directories. Do I have that right? It looks like disabling the enable_push_sync code path in device_utils.py's PushChangedFiles() also fixs this.

ah. that's a new code path that I'm working on for the emulator; pushing appears to be significantly faster, but it has some issues.

it would need to compare prior to calls to AdbWrapper.Push. DeviceUtils.PushChangedFiles already does a lot of this and is likely the layer in which a fix along these lines should be applied. given that it's a new code path that I'm working on, you're welcome to just disable enable_push_sync globally if you'd like & I can deal w/ fixing the issue.

Patch-set: 2
Reviewer: Gerrit User 1181690 <1181690@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>