Add global artifact logger
Adds a global artifact_logger module to Telemetry. This allows any part
of Telemetry to create artifacts for the current test without having
a reference to the object that's actually used to manage artifacts.

The specifics for each artifact implementation is abstracted behind
a compatability wrapper, so users don't have to care if Telemetry's or
typ's artifact implementation is used (or if an implementation has not
been set yet).

Bug: chromium:1008075
Change-Id: I8f024bd5d780d327d18e66a7df50544562531532