Implement timeout and retries for dev server tests
It's unfortunate when some component of dev server tests freezes and
we have to wait two hour for the bot timeout to be reached to see
the error. This CL adds:
- timeout-sec and timeout-retries flags to run_dev_server_tests.
Tests now have until timeout-sec to finish, and they can be retried
a number of times.
- Some timeouts for dashboard and tracing tests on build_steps. I have
seen these tests time out recently. The timeout value is set to
roughly max(duration on a passing run from 5 random samples of last 100
runs) * 2.

Bug: chromium:1015942
Change-Id: Ie5aecae2f94f3f9c0d23b6a499fa3cd49a64df05