Fix autocomplete-box to respect free-form user input.
The current implementation of autocomplete-box doesn't preserve
free-form input (i.e. typed rather than selected from the autocompletions).
So the field you see may differ from the field that is sent, confusing
users. This changes it to use the user input even if it doesn't match
any of the suggestions.

Bug: chromium:1016542
Change-Id: Ic6b5f29d80cd89879403fe07de6eff37ac6c0189