SheriffRedux: Generate Subscription from Sheriff
This change adds a facility available to administrators that will
translate the existing Sheriff entity into a Subscription protocol
buffer. These don't contain the anomaly configurations, but this allows
us to start moving the protocol buffers into configuration files.

Future changes will allow us to integrate the subscription
configurations incrementally, by adding the matching subscription
details in the test metadata and the alerts we generate (without acting
on the data or exposing the information yet).

For now this allows administrators to propose signing up some initial
users and setting up the files in version control somewhere.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:953181
Change-Id: Ib03719065ec1474804f752a4ffcda965461401f3