Pinpoint: Support jobs with no 'chart' argument
It turns out for try jobs we might not be providing the 'chart' argument
to the request, and we still end up trying to construct a path for
matching the histograms when reading the values.

Since try jobs will work anyway even if the histograms don't match
(because we'll generate a results2.html file that's iframe embedded) for
now, we update the histogram helper to be resilient to the case when the
histogram name is not provided while still having a story provided as
well.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1025229
Change-Id: Iebd0ad6a5825d06a4c3be9f9a3c95f1b093c5dfe