[Telemetry] Simplify Telemetry progress reporter
To make it more consistent with how test results are reported elswhere,
we directly use the story_run.test_path instead of trying to build its
own identifier.

The only visible effect is for the case of stories with grouping keys
which used to be written by the progress reporter, but wont be after
this change.

Grouping keys in the progress report were added long time ago by
https://chromium.googlesource.com/catapult.git/+/d32fc8b4320331ecca29ce180cba2b95ab21374c
to help distinguish stories with the same name but different grouping keys.

However, story sets are no longer allowed to have the duplicate stories with
the same name any more, c.f.:
https://cs.chromium.org/chromium/src/third_party/catapult/telemetry/telemetry/story/story_set.py?rcl=473984fca427a384fe9f037db299ea1be636e83d&l=99

This CL also removes a call to results.IterStoryRuns which we intend to
deprecate.

Bug: chromium:973837
Change-Id: Idb10fbf5476ecb0fc1458bc3d6a38c87ea9e9190