Update V8 runtime_stats to use different pattern for optimize background
OptimizeConcurrent and RecompileConcurrent are not very intuitive names
for the time spent on background for optimization. Concurrent compilation
has main thread component and background component so it gets confusing
to name only the background component as OptimizeConcurrent.
Currently, we never use OptimizeConcurrent so it is safe to drop it in
this cl. We can remove RecompileConcurrent once the V8 version that
renames RecompileConcurrent to OptimizeBackground rolls in.

Change-Id: I485878172f0eb6efb4df1eed92630ebb92ce29d1