telemetry: Increase timeout for trace collection.
Collection from go devices is very slow at the moment, exceeding the
existing 120 second timeout for large traces. This patch increases the
timeout.

We should be able to lower it again once we collect proto traces from
the devices instead: In that case, the trace is smaller and the device
has to perform much less computation (no JSON conversion) when
we read the trace data.

Bug: chromium:1026822
Change-Id: Ie197684a10e58b42f20479a51e73892b9bb33b7e