Update UmaMetric handling misaligned Bins
This is paired with a Chromium review which is building the metric.
https://chromium-review.googlesource.com/c/chromium/src/+/1867117/

When adding this we encountered an error in subractBins_

What can occur, is that when there are two events, the later
can have additional bins which were not reported during the
tracing. This was throwing an error.

This change updates uma_metric.html to not error out when there are
these extra bins. As well, if the result of subtractBins_ is that
there are no counts, the shift is set to 0, and no updates given to
the histogram. Instead of throwing an additional error.

TEST=ran vs samples from the change listed above (rendering.mobile)
additionalBinsPostTracingDoesNotCrash
Bug: chromium:1022036
Change-Id: I1f87a0669dbd1682b72044c5952a8ffd9520110a