[Telemetry] Refactor code for Telemetry "list" command.
The new code separates getting the benchmark information out to a
json serializable dict (which may be reused by external tools)
from actually listing the benchmarks in human readable form to stdout.

We also now allow listing benchmarks without having to specify
a --browser argument.

And a "description" for each story is provided in the json output.

Bug: chromium:1029042
Change-Id: Iea0befa0ad0dc954dc523d6cf021820692cc5754