telemetry: Increase timeout for trace collection further.
Follow-up from https://chromium-review.googlesource.com/c/catapult/+/1929096.

The timeout still seems to small, attempt increasing it some more.

We should be able to lower it again once we collect proto traces from
the devices instead: In that case, the trace is smaller and the device
has to perform much less computation (no JSON conversion) when
we read the trace data.

Bug: chromium:1029465
Change-Id: I68806223bf0508ab8ac29e88e0652cb077972df2