[dashboard] Skip identity test_path migration tasks
When doing a migration from an old_pattern to a new_pattern it is
sometimes possible for the job to match some old_test_path's with
identical corresponding new_test_path.

Furthermore, its dangerous when this happens. What should probably be
a no-op job, in fact ends up deleting all data for the test_path
entirely (since when migration is completed the "old" path is deleted).

This CL makes it so that these "identity" migrations are detected and
skipped over when creating renaming tasks.

Bug: chromium:1022380
Change-Id: I92c5217be3d59ad3b0b2957639eeb2a0b2ad32b1