Update patch set 12
Patch Set 12: Commit-Queue+1

Hi Folks,

There's a bit of new code going in here in multiple places, some necessary for correctness and some not quite. Unfortunately testing this isn't easy given the way the scheduler is set-up, we may need to do some "live" production iteration just for the execution engine.

Alternatively, we can make it so that the FIFO scheduler will be execution-engine-aware (by looking at the job to see whether it should use the execution engine) and kick off the evaluation engine and get out of the way as a first step. This will let us deploy the execution engine implementation in a different service subscribed to the topic that Swarming sends messages to.

For now this CL bakes the logic into the existing infrastructure, before we eventually migrate all the poll handling to use the execution engine as well. But we'll need to do that only after we have 100% coverage of the use-cases current encapsulated in the quests.

Thoughts?

Patch-set: 12
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1
Work-in-progress: false