[Telemetry] Allow benchmarks to specify TBMv3 metrics
This patch allows benchmarks authors to specify TBMv3 in addition to
existing TBMv2 benchmarks to compute on recorded traces for a story.

Authors would use the same as the current API, but can prefix
metric names with the corresponding version. For example:

  tbm_options.SetTimelineBasedMetrics([
      'v2:tracingMetric',
      'v3:dummy_metric',
      'cpuTimeMetric',  # Defaults to v2.
  ])

Note: metric computation is carried out by Results Processor, and
currently TBMv3 metrics will just be ignored if trace_processor_shell
is not available.

Bug: chromium:1012687
Change-Id: I83e3c5c4dc2a7e12e17b26648c8e51769d5e239a