Update patch set 1
Patch Set 1:

(1 comment)

This looks reasonable to me.

Could you add to the commit description that this just does simple merging not error handling/ deduplication and why that was chosen?

I defer to the telemetry owner on whether that should be handled here or if that is okay.

Patch-set: 1
Reviewer: Gerrit User 1290431 <1290431@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>