Use configuration-provided values
Instead of using the configuration, we use the information in the
`arguments` dict which is synthesised from the managed configuration
settings instead of those just directly input by the user. Because these
lookups are done before we perform the graph creation, we'll just reuse
that information instead of attempting to reimplement it for now.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1031405
Change-Id: I69c27d08757bd48f2f3deb41b4d0bcd8c35bb62a