Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (3 comments)
> 
> I don't plan to submit this CL of course, but can you look it over and let me know what you think needs to be changed before I do a cluster telemetry run to compare current lcp vs lcp if we don't invalidate images?

I'll restate that I'm not sure this approach works because invalidate is not necessarily the same as remove. Can we do a manual test:
1. Add an image, wait for it to load (should get an LCP candidate event)
2. Add another larger image, with ~infinite load time (does this get LCP invalidate event?)

If we get an invalidate event in this case, then this approach does not work, right?

Patch-set: 1
Reviewer: Gerrit User 1162220 <1162220@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>