Fix graph updates in performance bisection
When performing a graph update, we should include the edges in the
subgraph that do not exist yet. This change does the right thing, and in
the process we also extend the test cases to properly handle the
serialization of pending tasks.

We also reduce the logging detail in the debugging case, in favour of
logging indices and id's instead of actual dictionaries.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1034369, chromium:1034393
Change-Id: I562d65ecc460b4f47189fc3ff549740c9f4394f8