Fix graph refinement logic
This change ensures that when we're scaling attempts between two changes
that compare 'unknown' that we're growing the number of attempts for
both sides of the change. This avoids the potential case when both
changes have different numbers of attempts, skewing the measurement and
causing the bisection to never being able to grow the number of attempts
on one side of the comparison.

Also, instead of using relative numbers, use the new size of the
expected number of attempts to simplify the logic further.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1034393
Change-Id: Idde025699240ef04490930dad20019adae6eb899