Add proto definition for the new Histograms C++ API.
The new C++ API enables gtest binaries from Chromium and WebRTC to
write histogramset JSON natively. This will enable all dashboard users
to move to protos in the long run, which enables us to remove old
formats, conversion scripts and get a more efficient wire format.

This location is easy to access from the dashboard code later (this is
where all the other protos are) and can also easily be accessed by
the tracing/ code.

Bug: chromium:1029452
Change-Id: I35314cf941cbe36cb0452b173ef5772fe9d875b9