Execution Engine: Find Culprits and Completion
This change does an end-to-end test of all the parts of a successful
bisection, including growing the graph and exploring the space. The
contrived test case has significant differences in between all the
participant changes, so that we can ensure we're catching all of those
at the end of the process.

We also ensure that if we end up finding a culprit, that we are showing
where these are in the bug update.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1034393
Change-Id: I59d9cc6acfce983d579182f3de19aa422645c9e6