Pinpoint: Use 'user_data' for buildbot updates
This change fixes the update handling for buildbot requests, which use
the 'user_data' field instead of 'userdata' which Swarming uses. This
inconsistency might require us to use different update paths and topics
in the future, but for now consolidate into a single handler.

We also do another pass at removing calls to the FromDict function in
the Change module to remove unnecessary back-end API calls.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1034373, chromium:1039162
Change-Id: I346422a4c570f180b0fec221e140f6e48795ebb8