Fix buildbucket pub/sub handling
It turns out the buildbucket service will nicely provide the 'build'
status in the notification, which we can check whether a build is
completed or not. It also helpfully helps us determine whether a build
is scheduled, rather than just being stuck in 'pending'. In the future
we can improve the fidelity of the status display of Pinpoint tasks by
reflecting the status of the buildbucket/swarming tasks more clearly if
we're handling more granular state transitions from these services.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1039584
Change-Id: Ib647f71ced041dbdbe39fef8e0838e53512b52e4