Switch default anomaly detection algorithm
This change switches the default anomaly detection algorithm to use the
E-Divisive approximation implementation in the Dashboard. The change
allows us to keep studying the effects of the algorithm change with the
logging statements, switching the 'current' and 'alternate' algorithms.

Initial studies showed that we can reduce the amounts of alerts we're
generating by 30-45% just with the algorithm change, without sacrificing
fidelity nor accuracy of alert placement.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:922256
Change-Id: I887c387c469b8e84fd4184b33d36c9cc58548685