Avoid infinite loop in updateMajorMarkData
updateMajorMarkData spins in a loop adding values to an array of
doubles, executing curX += majorMarkDistanceWorld each time. If
majorMarkDistanceWorld is small enough relative to curX then this
addition is a NOP and the loop becomes infinite, leading to an
out-of-memory failure. This change adds a check to detect this
scenario and skip the loop.

Bug: chromium:1039830
Change-Id: I13503ca4c4e3ef3b2c696ae9b3c6d14e8b1c93c7