Reland "[Typ] Allow typ expectations parser to escape the '*' character"
This CL relands crrev.com/c/1978216

The CL that this CL tries to reland had an issue where we were only looking
for the longest glob that matched the test name. In this CL we will look
for the expectation with the longest glob which also has specifiers that
satisfy the system condition.

R=isamsonov@google.com
TBR=dpranke@chromium.org,johnchen@chromium.org

Bug: chromium:1038130
Bug: chromium:1036130

Change-Id: I79b14e12e599710861757c3f8786cfbe7cc97dc9