Split off upload processing to a different service
This change creates a different service, called 'upload-processing'
which allows us to isolate the background processing work for uploads
(histograms and chartjson points) from user-facing upload processing.
The isolation should help ensure uploads are processed in a timely
manner, helping with latency, and that background work can proceed
without interfering with user-facing traffic.

We can tune the limits of the services in later changes.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1035930
Change-Id: I13a810600ae169cf1901a788a68db329dfcc37ea